feat: changed from flask to the static web page
